fix: preserve quotations when compute
