+++ test/helpers.ts
@@ -1,3 +1,6 @@
+import fs from 'fs/promises'
+import path from 'path'
+
 const stripPattern = /^[ \t]*(?=[^\s]+)/mg
 
 export function heredoc(text: string) {
@@ -6,4 +9,9 @@ export function heredoc(text: string) {
   return indentLen > 0
     ? text.replace(indent, '').trimStart().replace(/ +?$/, '')
     : text
+}
+
+export async function readFixture(fname: string) {
+  const fpath = path.join(__dirname, 'fixture', fname)
+  return (await fs.readFile(fpath, 'utf-8')).replace(/\r\n/g, '\n')
 }
